Version 1 of our code!
To run it, open up the terminal/command prompt and:
- Compile the code using 'g++ pp1.cpp'
- Run the exe using 'a.exe' in Windows and './a.out' in  Linux.

This program checks for the passwd_file and master_passwd files in the same directory. If they exist, it extracts the master password from master_passwrd and allows the user to enter a password. If they don't exist, it lets the user register by creating these files. It also has empty functions for each of the functionalities we need to implement.

It should be good enough for a C++ 101 and a starting point for us.