Version 2 of our project 1 with sample inputs.
I have used the C++ program to generate a password file with 3 sample entries. In this version, we have a data structure (vector) to store the indexed domain names within the program, and a way to register new accounts - this first checks that the account doesn't already exist in the vector, and if it doesn't, adds an entry to the data structure, as well as the passwd_file. Minus the security aspect, the register_account function has been fully implemented.